% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{acc_over_feature}
\alias{acc_over_feature}
\title{Boxplot showing the accuracy of a metric against a selected feature}
\usage{
acc_over_feature(
  rdata,
  one_metric,
  feature,
  color_by,
  max_y = NA,
  rotation = 0,
  num_rotation = 0,
  signif_y = NA,
  comparison = NULL,
  AP = NULL,
  remove_NA = F
)
}
\arguments{
\item{rdata}{a dataframe}

\item{one_metric}{the y axis}

\item{feature}{the x axis}

\item{color_by}{fill the boxes}

\item{max_y}{manually cap the maximum of y axis in case outlier occurs}

\item{rotation}{rotation of x axis labels}

\item{num_rotation}{rotation of the labels for the number of samples for
each bar}

\item{comparison}{whether to compute Kruskal-Wallis test}
}
\description{
Boxplot showing the accuracy of a metric against a selected feature
}
