% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{add_impr}
\alias{add_impr}
\title{Append odds ratio}
\usage{
add_impr(rdata, acc_data, one_metric = "SSIM")
}
\arguments{
\item{rdata}{a dataframe to be appended with the odss ratio}

\item{acc_data}{a dataframe used to calculate the odds ratio. It must have
the column: \code{method} (which must have 2 types of methods--'proposed' and
'ZF'), \code{paper} for the name of each study}
}
\value{
a dataframe
}
\description{
Calculate the odds ratio of performance improvement on one
metric compared to zero filled data
}
