% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{all_metrics_cor}
\alias{all_metrics_cor}
\title{Pairwise correlation matrix}
\usage{
all_metrics_cor(all_data, metrics, metrics2 = NULL)
}
\arguments{
\item{all_data}{a dataframe}

\item{metrics}{a vector of metrics along the rows}

\item{metrics2}{a vector of metrics along the columns}
}
\description{
Pairwise correlation matrix
}
