% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{arrange_plots}
\alias{arrange_plots}
\title{Same function as \code{grid.arrange}, with the functionality of integrating
ComplexHeatmap and adding panel labels}
\usage{
arrange_plots(
  grob_list,
  save_path,
  grid_layout,
  panel_label = NULL,
  margin_width = 1,
  margin_height = 1,
  panel_spacing = 0.05,
  plot_width = 9,
  plot_height = 9,
  AP = NULL
)
}
\arguments{
\item{grob_list}{a list of plots. They can be ggplots, ComplexHeatmap,
data.frame, character, image array}

\item{save_path}{where the plots are saved}

\item{grid_layout, }{same as the argument in \code{grid.arrange}}

\item{margin_width}{the width of the margin in inches. The default is 0.79in
(2cm). Same unit in \code{margin_height}, \code{page_width}, \code{page_height}}

\item{plot_width}{the width of each subplot.}

\item{plot_height}{same concept as plot_width}
}
\description{
Same function as \code{grid.arrange}, with the functionality of integrating
ComplexHeatmap and adding panel labels
}
\references{
TBdev package
}
