% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_stat.R
\name{barplot_1var_overlap}
\alias{barplot_1var_overlap}
\title{Barplot of each study that contains multiple features}
\usage{
barplot_1var_overlap(xx, var1, var2, ID_col = NULL, ...)
}
\arguments{
\item{xx}{a dataframe}

\item{var1}{the variable for the y axis}

\item{var2}{the variable to fill the color}

\item{...}{arguments for \code{ggbar_1var}}
}
\value{
a ggplot bar plot orientated horizontally
}
\description{
Barplot of each study that contains multiple features
}
