% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_stat.R
\name{barplot_2var_label}
\alias{barplot_2var_label}
\title{Barplot with percentage labels}
\usage{
barplot_2var_label(dat, var1, var2, AP = NULL, ...)
}
\arguments{
\item{dat}{a dataframe}

\item{var1}{the x axis}

\item{var2}{the y axis}
}
\value{
a barplot with the labels of the percentage of the first category
over each bar, as well as the correlation of this percentage with \code{var1}
}
\description{
Barplot with percentage labels
}
