% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stat.R
\name{compare_2_group_per_metric}
\alias{compare_2_group_per_metric}
\title{y_val ~ x_val if x_val is categorical}
\usage{
compare_2_group_per_metric(com_df, x_val, y_val, ID)
}
\description{
y_val ~ x_val if x_val is categorical
}
