% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_stat.R
\name{compare_continuous_var_over_feature}
\alias{compare_continuous_var_over_feature}
\title{Boxplot for continuous variable}
\usage{
compare_continuous_var_over_feature(x, x_var, y_var, AP = NULL)
}
\arguments{
\item{x_var}{a discrete variable}

\item{y_var}{continuous variables}
}
\description{
Boxplot for continuous variable
}
