% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stat.R
\name{compare_multi_features_multi_metrics}
\alias{compare_multi_features_multi_metrics}
\title{Calculate p values for multiple variables}
\usage{
compare_multi_features_multi_metrics(
  com_df,
  x_vals,
  y_vals,
  padj_method = "hochberg"
)
}
\value{
a dataframe with the following column:
\code{x}: the independent variable
\code{y}: the dependent variable
\code{feature}: which group in \code{x}
\code{presence}: the value of \code{y} in that group
\code{absence}: the value of \code{y} outside that group
\code{presence_num}: number of samples in that group
\code{absence_num}: number of samples outside that group
\code{p}: unadjusted p value
\code{p_adj}: adjusted p value
}
\description{
This function calculates the p value between all possible pairs
of variables between those in \code{x_vals} and those in \code{y_vals}. The \code{y_vals}
must be numerical. If the \code{x_vals} is logical, Wilcoxon rank sum test
will be carried out. If the \code{x_vals} is numerical, p value will be that from
linear regression.
}
