% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{config}
\alias{config}
\title{Configurate aesthetic parameters}
\format{
An object of class \code{list} of length 12.
}
\usage{
config
}
\description{
a list containing:
\describe{
\item{fontsize}{the font size of all graphical elements except
\code{ggplot2::geom_text} and \code{ggrepel::geom_text_repel}}
\item{point_fontsize}{the font size of labelled text in the graph by
\code{geom_text} and \code{geom_text_repel}}
\item{legend_point_size}{the point size in legend symbol, if the value is
discrete}
\item{normal_shape}{shape of points, must be between 21 and 24}
\item{font_fam}{font family}
\item{npg_pal}{a vector of color palettes}
\item{transparency}{color transparency}
\item{shape_ops}{the shapes for dots if multiple shapes are required}
\item{color_vec}{a named vector of colors to plot for each year from
2017 to 2020}
}
}
\author{
Yutong Chen
}
\keyword{datasets}
