% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{custom_tick}
\alias{custom_tick}
\title{Make the x or y axis only show the min, middle and max points}
\usage{
custom_tick(
  vec = NULL,
  x_y = "y",
  more_prec = 3,
  min_prec = 1,
  num_out = 3,
  limits = NULL,
  ...
)
}
\description{
Make the x or y axis only show the min, middle and max points
}
