% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{Deeks_test}
\alias{Deeks_test}
\title{perform Deeks' test on a metric}
\usage{
Deeks_test(x, one_metric = "SSIM", AP = NULL)
}
\arguments{
\item{x}{a dataframe with the following column: \code{testing.number} for the
number of samples used in testing, \code{ID} for the ID of each study}
}
\description{
perform Deeks' test on a metric
}
