% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{load_all_studies}
\alias{load_all_studies}
\title{Combine csv files containing performance data}
\usage{
load_all_studies(root)
}
\arguments{
\item{root}{where the csv files are stored. Each csv must contain the
following columns: \code{acceleration}, \code{method}, \code{paper}, and columns for
different metrics.}
}
\value{
a list
}
\description{
This function appends the following columns:
\code{paper}: extracted from the csv filename
\code{dataset}: if not in the csv file, it will be 'dataset0'
\code{mask}: by uniting the \code{dataset} and \code{mask} columns
}
