% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{merge_all_studies}
\alias{merge_all_studies}
\title{Clean the merged dataset}
\usage{
merge_all_studies(all_list)
}
\value{
a dataframe
}
\description{
This function cleans the output from \code{load_all_studies}.
Firstly, 2 columns are created: \code{metric} and \code{value} by melting the
dataframes in each element of the input list.
}
