% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{metric_acc}
\alias{metric_acc}
\title{Plot the performance metrics of all studies}
\usage{
metric_acc(
  x,
  metric_types = c("PSNR", "SSIM", "NMSE"),
  metric_col = "metric",
  metric_val_col = "value",
  shape_col = "method",
  log_shape = T,
  shape_fil = c("proposed", "ZF"),
  num_col = NULL,
  color_by = "acceleration",
  AP = NULL
)
}
\arguments{
\item{x}{a dataframe containing the column \code{paper}}

\item{metric_types}{which metrics to plot}

\item{log_shape}{whether to take the log of the column \code{shape_col}}

\item{num_col}{number of columns to show}

\item{color_by}{which feature to color}
}
\description{
Plot the performance metrics of all studies
}
