% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metric_val}
\alias{metric_val}
\title{Metric performance data}
\format{
An object of class \code{data.frame} with 7289 rows and 6 columns.
}
\usage{
metric_val
}
\description{
a dataframe with the following columns:
\describe{
\item{method}{comparison method to evaluate the model performance.
If the method is the proposed study, it is labelled 'proposed'.
'ZF' means zero-filling}
\item{acceleration}{the acceleration ratio under which the metric
value is calculated}
\item{paper}{which paper the data is collected}
\item{dataset}{the testing dataset used to calculate the metric.
If unknown, it will be 'dataset0'}
\item{metric}{which metric}
\item{value}{the value for the metric of the method}
\item{}{}
}
}
\author{
Yutong Chen
}
\keyword{datasets}
