% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{npg_colors}
\alias{npg_colors}
\title{Define color scale}
\usage{
npg_colors(color_fill = T, color_vec = NULL, rever = F, AP = NULL)
}
\arguments{
\item{color_fill}{whether to color (FALSE) or to fill the dots (TRUE)}

\item{rever}{whether to reverse the legend order}
}
\description{
Define color scale
}
