% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_red.R
\name{pca_GMM}
\alias{pca_GMM}
\title{Plot PCA with the cluster results and the mean and variance of the clusters
as ellipses}
\usage{
pca_GMM(
  x,
  selected_features,
  GMM_mean = NULL,
  GMM_var = NULL,
  dims = c(1, 2),
  cluster_col = "cluster",
  AP = NULL,
  ...
)
}
\description{
need to load mclust first
}
