% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{plot_PSNR_SSIM}
\alias{plot_PSNR_SSIM}
\title{Scatterplot with raw PSNR and SSIM scores on the x and y axes respectively}
\usage{
plot_PSNR_SSIM(
  x_df,
  color_by,
  rotation = 0,
  label_text = T,
  metrics = c("PSNR", "SSIM"),
  log_color = F,
  AP = NULL
)
}
\arguments{
\item{x_df}{a dataframe containing the performance data}

\item{color_by}{which feature in \code{x_df} to color}

\item{label_text}{whether to label individual dots}
}
\description{
Scatterplot with raw PSNR and SSIM scores on the x and y axes respectively
}
