% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recode_dict}
\alias{recode_dict}
\title{Change the display names of features}
\format{
An object of class \code{data.frame} with 41 rows and 2 columns.
}
\usage{
recode_dict
}
\description{
a dataframe containing:
\describe{
\item{original}{the original labels}
\item{new}{the labels to display}
}
}
\author{
Yutong Chen
}
\keyword{datasets}
