% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{replace_feature}
\alias{replace_feature}
\title{Use a cleaner term to represent the names of model design features}
\usage{
replace_feature(xx, recode_dict)
}
\arguments{
\item{xx}{a string}

\item{recode_dict}{a dataframe with 2 columns: \code{original} and \code{new}}
}
\description{
Use a cleaner term to represent the names of model design features
}
