% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_stat.R
\name{resolve_overlap}
\alias{resolve_overlap}
\title{Flatten each list in a dataframe}
\usage{
resolve_overlap(x_df, feature, ID_col)
}
\arguments{
\item{x_df}{a dataframe}

\item{feature}{which column to extract the features}

\item{ID_col}{which column becomes the ID for the new features}
}
\value{
a named vector
}
\description{
Dataframes may contain such a column that contains strings, in
which items are joined by some delimiter. This makes it difficult to perform
analysis in individual items in the strings. This function creates a new
vector with each element containing one item in the strings. Each ittem has
an ID label from the original dataframe.
}
