% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{return_aes_param}
\alias{return_aes_param}
\title{Obtain aesthetic settings}
\usage{
return_aes_param(aes_param, load_format_conf = NULL)
}
\arguments{
\item{aes_param}{a named list of aesthetic settings}
}
\description{
If the supplied list is NULL, then the default settings will be
returned. Otherwise, the named fields in the supplied list will replace the
corresponding field in the default settings
}
\examples{
# change the fontsize and font family
return_aes_param (list (fontsize=12, font_fam='Arial') )
}
\references{
TBdev package
}
