% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_stat.R
\name{size_freq}
\alias{size_freq}
\title{Histogram of image sizes}
\usage{
size_freq(dat, size_col, color_by, bw = 100, AP = NULL)
}
\arguments{
\item{dat}{a dataframe}

\item{size_col}{which column has size information. The pixel size data is
extracted from the column. For example, if a stirng in that column is '256,
256', then the pixel size is $256 \times 256$}

\item{color_by}{which column contains color information
extracted}
}
\description{
Histogram of image sizes
}
